requirements.txt
Django==5.1.1
djangorestframework==3.17.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
gunicorn==22.0.0